+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Changes to the `dply` crate are documented in this file.
 
 ## 0.1.9 - 2023-06-28
 ### 🔧 Changed
-* Improve glimpse formmatting.
+* Improve glimpse formatting.
 ### 🐛 Fixed
 * repl: Fix completions for absolute and tilde paths.
 * repl: Keep all completion columns for big dataframes.